Move guild objects and rename event file
Reorganize object packages by moving soundboard and stage into objects/guild and update all affected imports. Rename gateway/payloads/events/eventtypes.go to registry.go and adjust references in gateway payloads and REST clients (gateway/payloads/events/guildsoundboard.go, gateway/payloads/events/stageinstances.go, rest/soundboard.go, rest/stage.go) to use the new paths.
