+++ gateway/payloads/events/registry.go
