+++ objects/guild/stage/stage.go
